refactor: Вносит правки в название папок и файлов для соответствия критериям
