+++ src/components/preview-offer-card/preview-offer-card.tsx
@@ -1,4 +1,3 @@
-
 function PreviewOfferCard(): JSX.Element {
   return (
     <article className="cities__card place-card">
